rewrite some components
